refactor clause.rs as module cdb
	modified:   src/assign/propagate.rs
	modified:   src/assign/stack.rs
	modified:   src/bin/splr.rs
	new file:   src/cdb/cid.rs
	new file:   src/cdb/clause.rs
	renamed:    src/clause.rs -> src/cdb/db.rs
	new file:   src/cdb/lbd.rs
	new file:   src/cdb/mod.rs
	new file:   src/cdb/watch.rs
	modified:   src/eliminate.rs
	modified:   src/lib.rs
	modified:   src/solver.rs
	modified:   src/state.rs
	modified:   src/types.rs
	modified:   src/validate.rs
