+++ src/cdb/clause.rs
@@ -1,0 +1,179 @@
+use {
+    crate::types::*,
+    std::{
+        cmp::Ordering,
+        fmt,
+        ops::{Index, IndexMut, Range, RangeFrom},
+        slice::Iter,
+    },
+};
+
+/// API for Clause, providing literal accessors.
+pub trait ClauseIF {
+    /// return true if it contains no literals; a clause after unit propagation.
+    fn is_empty(&self) -> bool;
+    /// return an iterator over its literals.
+    fn iter(&self) -> Iter<'_, Lit>;
+    /// return the number of literals.
+    fn len(&self) -> usize;
+}
+
+impl Default for Clause {
+    fn default() -> Clause {
+        Clause {
+            lits: vec![],
+            rank: 0,
+            search_from: 1,
+            reward: 0.0,
+            flags: Flag::empty(),
+        }
+    }
+}
+
+impl Index<usize> for Clause {
+    type Output = Lit;
+    #[inline]
+    fn index(&self, i: usize) -> &Lit {
+        unsafe { self.lits.get_unchecked(i) }
+    }
+}
+
+impl IndexMut<usize> for Clause {
+    #[inline]
+    fn index_mut(&mut self, i: usize) -> &mut Lit {
+        unsafe { self.lits.get_unchecked_mut(i) }
+    }
+}
+
+impl Index<Range<usize>> for Clause {
+    type Output = [Lit];
+    #[inline]
+    fn index(&self, r: Range<usize>) -> &[Lit] {
+        &self.lits[r]
+    }
+}
+
+impl Index<RangeFrom<usize>> for Clause {
+    type Output = [Lit];
+    #[inline]
+    fn index(&self, r: RangeFrom<usize>) -> &[Lit] {
+        &self.lits[r]
+    }
+}
+
+impl IndexMut<Range<usize>> for Clause {
+    #[inline]
+    fn index_mut(&mut self, r: Range<usize>) -> &mut [Lit] {
+        &mut self.lits[r]
+    }
+}
+
+impl IndexMut<RangeFrom<usize>> for Clause {
+    #[inline]
+    fn index_mut(&mut self, r: RangeFrom<usize>) -> &mut [Lit] {
+        &mut self.lits[r]
+    }
+}
+
+impl<'a> IntoIterator for &'a Clause {
+    type Item = &'a Lit;
+    type IntoIter = Iter<'a, Lit>;
+    fn into_iter(self) -> Self::IntoIter {
+        self.lits.iter()
+    }
+}
+
+impl<'a> IntoIterator for &'a mut Clause {
+    type Item = &'a Lit;
+    type IntoIter = Iter<'a, Lit>;
+    fn into_iter(self) -> Self::IntoIter {
+        self.lits.iter()
+    }
+}
+
+impl From<&Clause> for Vec<i32> {
+    fn from(c: &Clause) -> Vec<i32> {
+        c.lits.iter().map(|l| i32::from(*l)).collect::<Vec<i32>>()
+    }
+}
+
+impl ClauseIF for Clause {
+    fn is_empty(&self) -> bool {
+        self.lits.is_empty()
+    }
+    fn iter(&self) -> Iter<'_, Lit> {
+        self.lits.iter()
+    }
+    fn len(&self) -> usize {
+        self.lits.len()
+    }
+}
+
+impl FlagIF for Clause {
+    fn is(&self, flag: Flag) -> bool {
+        self.flags.contains(flag)
+    }
+    fn set(&mut self, f: Flag, b: bool) {
+        self.flags.set(f, b);
+    }
+    fn turn_off(&mut self, flag: Flag) {
+        self.flags.remove(flag);
+    }
+    fn turn_on(&mut self, flag: Flag) {
+        self.flags.insert(flag);
+    }
+}
+
+impl PartialEq for Clause {
+    fn eq(&self, other: &Clause) -> bool {
+        self == other
+    }
+}
+
+impl Eq for Clause {}
+
+impl PartialOrd for Clause {
+    fn partial_cmp(&self, other: &Clause) -> Option<Ordering> {
+        if self.rank < other.rank {
+            Some(Ordering::Less)
+        } else if other.rank < self.rank {
+            Some(Ordering::Greater)
+        } else if self.reward > other.reward {
+            Some(Ordering::Less)
+        } else if other.reward > self.reward {
+            Some(Ordering::Greater)
+        } else {
+            Some(Ordering::Equal)
+        }
+    }
+}
+
+impl Ord for Clause {
+    fn cmp(&self, other: &Clause) -> Ordering {
+        if self.rank < other.rank {
+            Ordering::Less
+        } else if other.rank > self.rank {
+            Ordering::Greater
+        } else if self.reward > other.reward {
+            Ordering::Less
+        } else if other.reward > self.reward {
+            Ordering::Greater
+        } else {
+            Ordering::Equal
+        }
+    }
+}
+
+impl fmt::Display for Clause {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        let st = |flag, mes| if self.is(flag) { mes } else { "" };
+        write!(
+            f,
+            "{{{:?}{}{}{}}}",
+            vec2int(&self.lits),
+            st(Flag::LEARNT, ", learnt"),
+            st(Flag::DEAD, ", dead"),
+            st(Flag::ENQUEUED, ", enqueued"),
+        )
+    }
+}
